fix tower name (#101)

+++ bedwars-api/src/main/java/com/tomkeuper/bedwars/api/arena/generator/IGenerator.java
@@ -84,7 +84,7 @@ public interface IGenerator {
     /**
      * Change item spawn delay. In seconds.
      */
-    void setDelay(int delay);
+    void setDelay(double delay);
 
     /**
      * Set how many items should the generator spawn at once.
@@ -165,12 +165,12 @@ public interface IGenerator {
     /**
      * Get spawn rate delay.
      */
-    int getDelay();
+    double getDelay();
 
     /**
      * Get seconds before next item spawn.
      */
-    int getNextSpawn();
+    double getNextSpawn();
 
     /**
      * Get the spawn limit of the generators.
@@ -181,8 +181,10 @@ public interface IGenerator {
 
     /**
      * Set the remaining time till the next item spawn.
+     *
+     * @param nextSpawn time to next spawn in seconds
      */
-    void setNextSpawn(int nextSpawn);
+    void setNextSpawn(double nextSpawn);
 
     /**
      * Should the dropped items be stacked?
